sudo -n, improved execute_command_with_results()
Return both result and return value from
execute_command_with_results(). This resulted in
all users of this function to need adjustment.

The modified behaviour was introduced to enable
calling sudo with -n (non-interactive) and handle
the case when it fails in a nice manner.

The only use of sudo in cmu so far is when reading logs
with journalctl on localhost; it's not essential
functionality, so if the user has not configured
their system to allow sudo without password, we can
just exclude that information instead of failing.

Thanks to Eero Tamminen for reporting this issue.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>